Add sample two phase commit
This commit adds a sample more complicated than the jug sample. It will
be used in the demonstration of the model verification. The components
in the sample are currently unused so the warning of them not being used
is temporarily suppressed. When the model verification implementation
completes, the suppression will be removed.